Fixed warning about eventually loosing significant digits.

+++ src/libxsmm_malloc.c
@@ -513,7 +513,7 @@ LIBXSMM_API_DEFINITION int libxsmm_malloc_attrib(void** memory, int flags, const
     assert((0 != buffer || 0 == size) && 0 != internal);
     /* quietly keep the read permission, but eventually revoke write permissions */
     if (0 == (LIBXSMM_MALLOC_FLAG_W & flags) || 0 != (LIBXSMM_MALLOC_FLAG_X & flags)) {
-      const int alignment = ((const char*)(*memory)) - ((const char*)buffer);
+      const int alignment = (int)(((const char*)(*memory)) - ((const char*)buffer));
       const size_t alloc_size = size + alignment;
       /* treat mprotect errors as soft error if the requested buffer is not executable */
       int soft_error = 0;
